debug option to keep intermediate files

+++ variantiq-pipelines/src/pipelines/variantiq-bwa-bcftools.pipeline.sh
@@ -3,6 +3,7 @@ set -e
 
 PIPELINE_NAME="variantiq-bwa-bcftools"
 SAMPLE_FOLDER=$1
+DEBUG=$2
 REFERENCE_GENOME=${SAMPLE_FOLDER}/reference_genome.fasta
 READ1=${SAMPLE_FOLDER}/read_1.fastq
 READ2=${SAMPLE_FOLDER}/read_2.fastq
@@ -48,7 +49,9 @@ if [ ! -f "${SAMPLE_FOLDER}/${PIPELINE_NAME}.vcf" ]; then
 
     mv ${BUILD_FOLDER}/bcf_normalized_variants.vcf ${SAMPLE_FOLDER}/${PIPELINE_NAME}.vcf
 
-    rm -rf $BUILD_FOLDER
+    if [ -z "$DEBUG"]; then
+        rm -rf $BUILD_FOLDER
+    fi
 else
     echo "Skipping ${SAMPLE_FOLDER}/${PIPELINE_NAME}.vcf.  Already exists."
 fi
